refactor(Table): move interface types to index.tsx
